+++ packages/browser-destinations/src/destinations/adobe-target/trackEvent/index.ts
@@ -2,7 +2,7 @@ import type { BrowserActionDefinition } from '../../../lib/browser-destinations'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import { Adobe } from '../types'
-import { setMbox3rdPartyId } from '../utils'
+import { setMbox3rdPartyId, serializeProperties } from '../utils'
 
 // Adobe Target only takes certain event types as valid parameters. We are defaulting to "click".
 const TARGET_EVENT_TYPE = 'click'
@@ -53,26 +53,8 @@ const action: BrowserActionDefinition<Settings, Adobe, Payload> = {
     const payload = event.payload
     setMbox3rdPartyId(payload.userId)
 
-    // Track does not accept arrays as valid properties, therefore we are stringifying them.
-    const serialize_properties = (props: { [key: string]: unknown } | undefined) => {
-      if (props === undefined) {
-        return {}
-      }
-
-      const serialized: { [key: string]: unknown } = {}
-
-      for (const key in props) {
-        serialized[key] = props[key]
-        if (Array.isArray(props[key])) {
-          serialized[key] = JSON.stringify(props[key])
-        }
-      }
-
-      return serialized
-    }
-
     const event_params = {
-      ...serialize_properties(payload.properties),
+      ...serializeProperties(payload.properties),
       event_name: payload.eventName
     }
